+++ tests/integration/test_hierarchy_to_classifier.py
@@ -19,8 +19,8 @@ import polars as pl
 
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.domain.enums import ApproachType, ExposureClass
-from rwa_calc.engine.classifier import ExposureClassifier
-from rwa_calc.engine.hierarchy import HierarchyResolver
+from rwa_calc.engine.stages.classify import ExposureClassifier
+from rwa_calc.engine.stages.hierarchy import HierarchyResolver
 
 from .conftest import (
     make_contingent,